Add zones and saving logic
